PostFetcher: replace JAXB with JSoup for html parsing.
SAXParser blows up when it reaches an entity, e.g. &raqou;, that is not
declared. I mocked everything and though I couldn't avoid that. I mocked
XMLReader and passed it to the SAXSource that I pass to JAXB's Unmarshaller, so
that if Unmarshaller sets ErrorHandler or ContentHandler to it I will mock them
(all methods in ContentHandler were mocked). It turns out that the error is
thrown by XMLReader.parse(). The error, 'entity raquo is reference but not
declared', is fatal and even though I said ErroHanldler not to throw Exception,
parse() throws one on behalf of himself. This seems to me contradicting with
Oracle's description about the role of ErrorHandler in XMLReader. I use Java 8,
btw.

EntityResolver is also set but it is not called ever.

It turns out that it is a wrong practice to parse html documents with a heavily
syntax-validating xml parser, like SAXParser.